Revert timing code and move to System/Timing/X
