+++ Source/SysPSP/HLEGraphics/GraphicsPluginPSP.cpp
@@ -33,7 +33,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "Utility/Profiler.h"
 #include "Utility/FramerateLimiter.h"
 #include "Interface/Preferences.h"
-#include "System/Timing.h"
+#include "System/Timing/Timing.h"
 
 #include <pspdebug.h>
