+++ Source/Utility/BatchTest.cpp
@@ -40,7 +40,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "RomFile/RomFile.h"
 #include "Utility/Timer.h"
-#include "System/Timing.h"
+#include "System/Timing/Timing.h"
 
 void MakeRomList( const std::filesystem::path& romdir, std::vector< std::string > & roms )
 {
