+++ Source/Utility/FramerateLimiter.cpp
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "Base/Types.h"
 #include "FramerateLimiter.h"
 
-#include "System/Timing.h"
+#include "System/Timing/Timing.h"
 #include "System/Thread.h"
 
 #include "Core/Memory.h"
