+++ Source/Utility/Profiler.cpp
@@ -18,7 +18,7 @@
 #include "Debug/DBGConsole.h"
 #include "Utility/Hash.h"
 #include "Utility/Profiler.h"
-#include "System/Timing.h"
+#include "System/Timing/Timing.h"
