+++ Source/Utility/Timer.cpp
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "Base/Types.h"
 #include "Utility/Timer.h"
 
-#include "System/Timing.h"
+#include "System/Timing/Timing.h"
 
 //*************************************************************************************
 //
